% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_data.R
\docType{data}
\name{airport_data}
\alias{airport_data}
\title{Latitude/Longitude Coordinates of Airports}
\format{
a dataframe with '354' rows and '3' columns
\describe{
\item{codes}{name of airport}
\item{lon}{longitude of airport, in decimal degrees}
\item{lat}{latitude of airport, in decimal degrees}
}
}
\usage{
airport_data
}
\description{
A dataset with the coordinates of 'x' airports
}
\keyword{datasets}
