% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airports_nearby.R
\name{airports_nearby}
\alias{airports_nearby}
\title{Names of Airports within 'x' Miles of given lon lat coordinates}
\usage{
airports_nearby(
  lonlat,
  lonlat_code,
  lonlat_airport_total,
  lonlat_airport_codes,
  x = 75
)
}
\arguments{
\item{lonlat}{a 1 x 2 matrix with columns 'lon' 'lat', values should be in decimal degrees, (positive for North and East, negative for South and West)}

\item{lonlat_code}{a string with the airport code associated with lonlat}

\item{lonlat_airport_total}{a 'n' x 2 matrix with columns 'lon' 'lat', for all airports to be included in the search}

\item{lonlat_airport_codes}{a vector of length 'n' with the airport names associated with, the lonlat pairs included in lonlat_airport_total}

\item{x}{the maximum distance an airport can be from given lonlat coordinate pair}
}
\description{
Gets 3-digit airport code of airports within 'x' miles of given point
}
\examples{
# get airports within 75 miles of Ithaca, NY
ith = airport_data[which(airport_data$codes == 'ITH'),]
ll1 = data.frame(lon = ith[1,2], lat = ith[1,3])
ll1_codes = ith$codes[1]
ll_airport_total = data.frame(lon = airport_data[,2], lat = airport_data[,3])
ll_airport_codes = airport_data$codes
output = airports_nearby(ll1, ll1_codes, ll_airport_total, ll_airport_codes)

}
