% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_symmetric.R
\name{check_symmetric}
\alias{check_symmetric}
\title{Function that returns whether direct_from object is symmetric}
\usage{
check_symmetric(direct_from)
}
\arguments{
\item{direct_from}{A list of character vectors, indexable via airport code
 \itemize{
  \item \code{symm} Boolean that indicates whether direct_from is symmetric
  \item \code{conn} n x 2 character matrix that provides the 'n' asymmetric connections
}}
}
\description{
Symmetry in direct_from indicates that there for any airports A and B, A has a direct flight to B and B has a direct flight to A
}
\examples{
# check whether default direct_from object is symmetric
# there are no asymmetric connections
check_symmetric(direct_from)

}
