% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_from.R
\docType{data}
\name{direct_from}
\alias{direct_from}
\title{list containing information about direct connections from each airport}
\format{
a list of length 354, names are 3-letter airport codes
\describe{
\item{vector}{direct connections from each airport}
}
}
\usage{
direct_from
}
\description{
The jth list element of the list should contain a character vector of the
nonstop destinations reachable from the jth airport in airport_data
}
\keyword{datasets}
