% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_from_a_100.R
\docType{data}
\name{direct_from_a_100}
\alias{direct_from_a_100}
\title{List containing information about direct connections from each airport under redesigned network A, 100 mile radius}
\format{
a list of length 354, names are 3-letter airport codes
\describe{
\item{vector}{direct connections from each airport}
}
}
\usage{
direct_from_a_100
}
\description{
List containing information about direct connections from each airport under redesigned network A, 100 mile radius
}
\keyword{datasets}
