% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lonlat_usa.R
\docType{data}
\name{lonlat_usa}
\alias{lonlat_usa}
\title{Latitude/Longitude Coordinates of Continental USA Gridpoints}
\format{
a dataframe with '9063' rows and '2' columns
\describe{
\item{lon}{longitude of grid point, in decimal degrees}
\item{lat}{latitude of grid point, in decimal degrees}
}
}
\usage{
lonlat_usa
}
\description{
A dataset of lonlat pairs that are representative of the continental USA
}
\keyword{datasets}
