% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redesign_A100.R
\docType{data}
\name{redesign_A100}
\alias{redesign_A100}
\title{Redesigned network under scenario A, 100 mile radius}
\format{
a list of length 2
\describe{
\item{direct_to}{direct_to list object for redesigned network}
\item{prop}{proportion of reachable continental US gridpoints}
}
}
\usage{
redesign_A100
}
\description{
Redesigned network under scenario A, 100 mile radius
}
\keyword{datasets}
