% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redesign_C100.R
\docType{data}
\name{redesign_C100}
\alias{redesign_C100}
\title{Redesigned network under scenario C, 100 mile radius}
\format{
a list of length 2
\describe{
\item{direct_to}{direct_to list object for redesigned network}
\item{prop}{proportion of reachable continental US gridpoints}
}
}
\usage{
redesign_C100
}
\description{
Redesigned network under scenario C, 100 mile radius
}
\keyword{datasets}
